Rename namespace "X4_ComplexCalculator.Main.Menu.Window" ↓↓↓ "X4_ComplexCalculator.Main.Menu.View"
"X4_ComplexCalculator.Main.WorkArea.UI.Menu.View"
↓↓↓
"X4_ComplexCalculator.Main.WorkArea.UI.Menu.Tab"
