+++ X4_ComplexCalculator/Localization/Lang.csv
@@ -62,8 +62,8 @@ DBUpdateConfirmationMessage;Do you want to display the DB update screen? \r\nPle
 DBUpdateRestartRequestMessage;Please restart the application to apply the DB update.
 DBUpdateFailureMessage;Failed to update the DB, \r\nplease check if the DB file is accessible and try again.
 
-# Window(W) --------------------------------------------
-Window;Window(_W)
+# View(W) ----------------------------------------------
+View;View(_V)
 
 # Empire overview --------------------------------------
 EmpireOverview;Empire overview
@@ -141,7 +141,7 @@ PresetImported;Imported.
 # ------------------------------------------------------
 # Plan
 # ------------------------------------------------------
-View;View(_V)
+Tab;Tab
 ModuleList;Modules
 Products;Products
 Summary;Summary
