+++ X4_ComplexCalculator/Localization/Lang.ja-JP.csv
@@ -62,8 +62,8 @@ DBUpdateConfirmationMessage;DB更新画面を表示しますか？\r\n※ 画面
 DBUpdateRestartRequestMessage;DBの更新を適用するためには本アプリケーションの再起動を行ってください。
 DBUpdateFailureMessage;DBの更新に失敗しました。\r\nDBファイルにアクセス可能か確認後、再度実行してください。
 
-# Window(W) --------------------------------------------
-Window;ウィンドウ(_W)
+# View(W) ----------------------------------------------
+View;表示(_V)
 
 # Empire overview --------------------------------------
 EmpireOverview;帝国の概要
@@ -135,7 +135,7 @@ PresetImported;インポート済み
 # ------------------------------------------------------
 # Plan
 # ------------------------------------------------------
-View;表示(_V)
+Tab;タブ
 ModuleList;モジュール
 Products;製品
 Summary;概要
