+++ X4_ComplexCalculator/Main/MainWindow.xaml
@@ -126,8 +126,8 @@
             </MenuItem>
 
 
-            <!--  ウィンドウ  -->
-            <MenuItem Header="{lex:Loc Lang:Window}">
+            <!--  表示(V)  -->
+            <MenuItem Header="{lex:Loc Lang:View}">
                 <!--  帝国の概要  -->
                 <MenuItem Command="{Binding OpenEmpireOverviewWindowCommand}" Header="{lex:Loc Lang:EmpireOverview}" />
