+++ X4_ComplexCalculator/Main/MainWindowViewModel.cs
@@ -23,7 +23,7 @@ using X4_ComplexCalculator.Main.Menu.File.Import.LoadoutImport;
 using X4_ComplexCalculator.Main.Menu.File.Import.StationPlanImport;
 using X4_ComplexCalculator.Main.Menu.Lang;
 using X4_ComplexCalculator.Main.Menu.Layout;
-using X4_ComplexCalculator.Main.Menu.Window.EmpireOverview;
+using X4_ComplexCalculator.Main.Menu.View.EmpireOverview;
 using X4_ComplexCalculator.Main.WorkArea;
 
 namespace X4_ComplexCalculator.Main
