+++ X4_ComplexCalculator/Main/Menu/File/Import/LoadoutImport/SelectLoadoutDialog.xaml.cs
@@ -7,7 +7,7 @@ namespace X4_ComplexCalculator.Main.Menu.File.Import.LoadoutImport
     /// <summary>
     /// SelectLoadoutDialog.xaml の相互作用ロジック
     /// </summary>
-    public partial class SelectLoadoutDialog : System.Windows.Window
+    public partial class SelectLoadoutDialog : Window
     {
         /// <summary>
         /// コンストラクタ
@@ -25,7 +25,7 @@ namespace X4_ComplexCalculator.Main.Menu.File.Import.LoadoutImport
         {
             var wnd = new SelectLoadoutDialog();
 
-            wnd.Owner = Application.Current.Windows.OfType<System.Windows.Window>().FirstOrDefault(x => x.IsActive) ?? Application.Current.MainWindow;
+            wnd.Owner = Application.Current.Windows.OfType<Window>().FirstOrDefault(x => x.IsActive) ?? Application.Current.MainWindow;
 
             return wnd.ShowDialog() == true;
         }
