+++ X4_ComplexCalculator/Main/Menu/File/Import/SaveDataImport/SelectStationDialog.xaml.cs
@@ -7,7 +7,7 @@ namespace X4_ComplexCalculator.Main.Menu.File.Import.SaveDataImport
     /// <summary>
     /// SelectStationDialog.xaml の相互作用ロジック
     /// </summary>
-    public partial class SelectStationDialog : System.Windows.Window
+    public partial class SelectStationDialog : Window
     {
         /// <summary>
         /// コンストラクタ
@@ -28,7 +28,7 @@ namespace X4_ComplexCalculator.Main.Menu.File.Import.SaveDataImport
         {
             var wnd = new SelectStationDialog(stationItems);
 
-            wnd.Owner = Application.Current.Windows.OfType<System.Windows.Window>().FirstOrDefault(x => x.IsActive) ?? Application.Current.MainWindow;
+            wnd.Owner = Application.Current.Windows.OfType<Window>().FirstOrDefault(x => x.IsActive) ?? Application.Current.MainWindow;
 
             return wnd.ShowDialog() == true;
         }
