+++ X4_ComplexCalculator/Main/Menu/View/EmpireOverview/EmpireOverviewWindow.xaml
@@ -1,11 +1,11 @@
 ﻿<Window
-    x:Class="X4_ComplexCalculator.Main.Menu.Window.EmpireOverview.EmpireOverviewWindow"
+    x:Class="X4_ComplexCalculator.Main.Menu.View.EmpireOverview.EmpireOverviewWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:i="http://schemas.microsoft.com/xaml/behaviors"
     xmlns:lex="http://wpflocalizeextension.codeplex.com"
-    xmlns:local="clr-namespace:X4_ComplexCalculator.Main.Menu.Window.EmpireOverview"
+    xmlns:local="clr-namespace:X4_ComplexCalculator.Main.Menu.View.EmpireOverview"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     Title="{lex:Loc Lang:EmpireOverview}"
     Width="400"
