+++ X4_ComplexCalculator/Main/WorkArea/UI/Menu/Tab/VisiblityMenuItem.cs
@@ -2,7 +2,7 @@
 using AvalonDock.Layout;
 using Prism.Mvvm;
 
-namespace X4_ComplexCalculator.Main.WorkArea.UI.Menu.View
+namespace X4_ComplexCalculator.Main.WorkArea.UI.Menu.Tab
 {
     /// <summary>
     /// 表示状態メニュー1レコード分のクラス
