+++ X4_ComplexCalculator/Main/WorkArea/WorkAreaViewModel.cs
@@ -17,7 +17,7 @@ using X4_ComplexCalculator.DB;
 using X4_ComplexCalculator.Main.Menu.File.Export;
 using X4_ComplexCalculator.Main.Menu.File.Import;
 using X4_ComplexCalculator.Main.WorkArea.UI.BuildResourcesGrid;
-using X4_ComplexCalculator.Main.WorkArea.UI.Menu.View;
+using X4_ComplexCalculator.Main.WorkArea.UI.Menu.Tab;
 using X4_ComplexCalculator.Main.WorkArea.UI.ModulesGrid;
 using X4_ComplexCalculator.Main.WorkArea.UI.ProductsGrid;
 using X4_ComplexCalculator.Main.WorkArea.UI.StationSummary;
